Reggie_Online/ApiSync/backend/requirements.txt
# ApiSync - Python Dependencies
# FastAPI web framework and ASGI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Google Cloud Firestore
google-cloud-firestore>=2.13.0

# Environment variable management
python-dotenv>=1.0.0

# Data validation (required by FastAPI)
pydantic>=2.0.0

# Async HTTP client for external API calls
httpx>=0.25.0

